Remove unused files and reorganize existing ones into appropriate folders
