+++ src/index.css
@@ -6,7 +6,7 @@ body {
     'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-image: url('images/background.jpg');
+  background-image: url('./assets/images/background.jpg');
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
